update type of AI feedback

+++ core/infrastructure/src/main/java/org/naho/vocabulary/adapter/VocabularyRepositoryAdapter.java
@@ -2,8 +2,8 @@ package org.naho.vocabulary.adapter;
 
 import lombok.RequiredArgsConstructor;
 import org.naho.question.entity.VocabularyQuestionEntity;
+import org.naho.question.model.Vocabulary;
 import org.naho.question.repository.VocabularyQuestionJpaRepository;
-import org.naho.vocabulary.model.Vocabulary;
 import org.naho.vocabulary.port.out.VocabularyPort;
 import org.springframework.stereotype.Component;
 
@@ -17,8 +17,8 @@ public class VocabularyRepositoryAdapter implements VocabularyPort {
     private final VocabularyQuestionJpaRepository vocabularyQuestionJpaRepository;
 
     @Override
-    public List<Vocabulary> findVocabularyList(int vocabularyQuestionId) {
-        return vocabularyQuestionJpaRepository.findById((long) vocabularyQuestionId)
+    public List<Vocabulary> findVocabularyList(Long vocabularyQuestionId) {
+        return vocabularyQuestionJpaRepository.findById(vocabularyQuestionId)
                 .map(VocabularyQuestionEntity::getVocabularies)
                 .map(entities -> entities.stream()
                         .map(entity -> Vocabulary.builder()
